// src/screens/products/types.ts
export type TProductCreate = {
  product_type: string;
  book: string;
  product_name: string;
  calypso_le: string;
  ref_ccy: string;
  gl: string;
  spot: string;
  show_forma: any;
  show_dorm: any;
  show_allocation: any;
};
export type TProductEdit = {};
export type TProductView = {};
export type TProductList = {};